requirements.txt
google-generativeai
python-dotenv
speechrecognition
sounddevice
scipy
fer
opencv-python
transformers
gtts
googletrans==4.0.0-rc1
pyttsx3
pyaudio
